General Social Survey, 2024

These data come from the General Social Survey, administered in 2024. These data were collected by the National Opinion Research Center (NORC) at the University of Chicago. The observations represent individuals' responses to survey questions. The full data set – The General Social Survey Cumulative Data (1972-2024, release 3) – was generously provided through the gssr package. Further information about the GSS2024 data set (including question text) can be found in the GSS 2024 Codebook at: https://gss.norc.org/content/dam/gss/get-documentation/pdf/codebook/GSS%202024%20Codebook%20R3.pdf.

Usage

GSS2024

Format

A data frame with 3986 observations and 983 variables.

year

GSS year for this respondent	id

Respondent id number	wrkstat

labor force status	hrs1

number of hours worked last week	hrs2

number of hours usually work a week	evwork

ever work as long as one year	wrkslf

r self-emp or works for somebody	occ10

r's census occupation code (2010)	prestg10

r's occupational prestige score (2010)	prestg105plus

R's occupational prestige score using threshold method (2010)	indus10

r's industry code (naics 2007)	marital

marital status	martype

marital type	divorce

ever been divorced or separated	widowed

ever been widowed	spwrksta

spouse labor force status	sphrs1

number of hrs spouse worked last week	sphrs2

no. of hrs spouse usually works a week	spevwork

spouse ever work as long as a year	cowrksta

partner's labor force status	cowrkslf

partner self-emp. or works for somebody else	coevwork

partner ever work as long as a year	cohrs1

number of hours partner worked last week	cohrs2

no. of hours partner usually works a week	spwrkslf

spouse self-emp. or works for somebody	spocc10

spouse census occupation code (2010)	sppres10

Spouse occupational prestige score (2010)	sppres105plus

Spouse occupational prestige score using threshold method (2010)	spind10

spouse's industry code (naics 2007)	coocc10

partner's census occupation code (2010)	coind10

partner's industry code (2010)	pawrkslf

Father (or oldest same-sex parent) self-emp. or worked for somebody	paocc10

father's census occupation code (2010)	papres10

father's occupational prestige score (2010)	papres105plus

Father's occupational prestige score using threshold method (2010)	paind10

father's industry code (naics 2007)	mawrkslf

Mother (or youngest same-sex parent) self-emp. or worked for somebody	maocc10

mothers census occupation code (2010)	mapres10

mothers occupational prestige score (2010)	mapres105plus

Mothers occupational prestige score using threshold method (2010)	maind10

mothers industry code (naics 2007)	sibs

number of brothers and sisters	childs

number of children	age

age of respondent	agekdbrn

r's age when 1st child born	educ

highest year of school completed	paeduc

Highest year school completed, father (or oldest same-sex parent)	maeduc

Highest year school completed, mother (or youngest same-sex parent)	speduc

highest year school completed, spouse	coeduc

highest year school completed, partner	codeg

partner's highest degree	degree

r's highest degree	padeg

Father's (or oldest same-sex parent's) highest degree	madeg

Mother's (or youngest same-sex parent's) highest degree	spdeg

spouse's highest degree	major1

college major 1	major2

college major 2	dipged

diploma, ged, or other	sex

respondents sex	race

race of respondent	res16

type of place lived in when 16 yrs old	reg16

region of residence, age 16	mobile16

geographic mobility since age 16	family16

living with parents when 16 yrs old	famdif16

reason not living with parents	mawrkgrw

Did mother (or youngest same-sex parent) ever work for pay for as long as a year while you were growing up?	incom16

r's family income when 16 yrs old	born

was r born in this country	parborn

were r's parents born in this country	granborn

how many grandparents born outside u.s.	hompop

number of persons in household	babies

household members less than 6 yrs old	preteen

household members 6 thru 12 yrs old	teens

household members 13 thru 17 yrs old	adults

household members 18 yrs and older	unrelat

number in household not related	earnrs

how many in family earned money	income

total family income	rincome

respondents income	income16

total family income	rincom16

respondents income	region

region of interview	xnorcsiz

expanded norc size code	partyid

political party affiliation	vote16

remember if voted in 2016 election	pres16

vote clinton or trump	if16who

who you would have voted for	polviews

think of self as liberal or conservative	natspac

space exploration program	natenvir

improving & protecting environment	natheal

improving & protecting nations health	natcity

solving problems of big cities	natcrime

halting rising crime rate	natdrug

dealing with drug addiction	nateduc

improving nations education system	natrace

Improving the conditions of Black people	natarms

military, armaments, and defense	nataid

foreign aid	natfare

welfare	natroad

highways and bridges	natsoc

social security	natchld

assistance for childcare	natsci

supporting scientific research	natenrgy

developing alternative energy sources	natspacy

space exploration:version y	natenviy

the environment:version y	nathealy

health:version y	natcityy

assistance to big cities:version y	natcrimy

law enforcement:version y	natdrugy

drug rehabilitation:version y	nateducy

education:version y	natracey

Assistance to Black people: version y	natarmsy

national defense:version y	nataidy

assistance to other countries:ver y	natfarey

assistance to the poor:version y	eqwlth

should govt reduce income differences	tax

r's federal income tax	spkath

allow anti-religionist to speak	colath

allow anti-religionist to teach	libath

allow anti-religious book in library	spkrac

allow racist to speak	colrac

allow racist to teach	librac

allow racists book in library	spkcom

allow communist to speak	colcom

should communist teacher be fired	libcom

allow communists book in library	spkmslm

allow muslim clergymen preaching hatred of the us	colmslm

allow anti-american muslim clergymen teaching in college	libmslm

allow anti-american muslim clergymen's books in library	cappun

favor or oppose death penalty for murder	gunlaw

favor or oppose gun permits	courts

courts dealing with criminals	grass

should marijuana be made legal	relig

r's religious preference	denom

specific denomination	fund

how fundamentalist is r currently	attend

how often r attends religious services	reliten

strength of affiliation	postlife

belief in life after death	pray

how often does r pray	relig16

religion in which raised	denom16

denomination in which r was raised	fund16

how fundamentalist was r at age 16	sprel

spouse's religious preference	spden

specific denomination, spouse	spfund

how fundamentalist is spouse currently	corel

partner's religious preference	coden

partner's denomination	cofund

how fundamentalist is partner's current religion	prayer

bible prayer in public schools	bible

feelings about the bible	racopen

vote on open housing law	raclive

any opp. race in neighborhood	affrmact

Favor preference in hiring Black people	wrkwayup

Black people overcome prejudice without favors	happy

general happiness	hapmar

happiness of marriage	hapcohab

happiness of relt with partner	health

condition of health	life

is life exciting or dull	helpful

people helpful or looking out for selves	fair

people fair or try to take advantage	trust

can people be trusted	confinan

confid in banks & financial institutions	conbus

confidence in major companies	conclerg

confidence in organized religion	coneduc

confidence in education	confed

confid. in exec branch of fed govt	conlabor

confidence in organized labor	conpress

confidence in press	conmedic

confidence in medicine	contv

confidence in television	conjudge

confid. in united states supreme court	consci

confidence in scientific community	conlegis

confidence in congress	conarmy

confidence in military	obey

to obey	popular

to be well liked or popular	thnkself

to think for ones self	workhard

to work hard	helpoth

to help others	socrel

spend evening with relatives	socommun

spend evening with neighbor	socfrend

spend evening with friends	socbar

spend evening at bar	weekswrk

weeks r worked last year	partfull

was r's work part-time or full-time?	joblose

is r likely to lose job	jobfind

could r find equally good job	satjob

work satisfaction	richwork

if rich, continue or stop working	class

subjective class identification	rank

r's self ranking of social position	satfin

satisfaction with financial situation	finalter

change in financial situation	finrela

opinion of family income	wksub

does r or spouse have supervisor	wksubs

does supervisor have supervisor	wksub1

does r or spouse or partner have a supervisor	wksubs1

does supervisor have supervisor	wksup

does r or spouse supervise anyone	wksups

does subordinate supervise anyone	wksup1

does r or spouse or partner supervise anyone	wksups1

does subordinant supervise anyone	unemp

ever unemployed in last ten yrs	union

does r or spouse belong to union	union1

does r or spouse or partner belong to union	getahead

opinion of how people get ahead	parsol

r's living standard compared to parents	kidssol

r's kids living standard compared to r	fepol

women not suited for politics	abdefect

strong chance of serious defect	abnomore

married–wants no more children	abhlth

woman's health seriously endangered	abpoor

low income–cant afford more children	abrape

pregnant as result of rape	absingle

not married	abany

abortion if woman wants for any reason	chldidel

ideal number of children	pillok

birth control to teenagers 14-16	sexeduc

sex education in public schools	divlaw

divorce laws	premarsx

sex before marriage	teensex

sex before marriage:teens 14-16	xmarsex

sex with person other than spouse	homosex

homosexual sex relations	pornlaw

feelings about pornography laws	xmovie

seen x-rated movie in last year	spanking

favor spanking to discipline child	letdie1

allow incurable patients to die	suicide1

suicide if incurable disease	suicide2

suicide if bankrupt	suicide4

suicide if tired of living	polhitok

ever approve of police striking citizen	polabuse

citizen said vulgar or obscene things	polmurdr

citizen questioned as murder suspect	polescap

citizen attempting to escape custody	polattak

citizen attacking policeman with fists	fear

afraid to walk at night in neighborhood	owngun

have gun in home	pistol

pistol or revolver in home	shotgun

shotgun in home	rifle

rifle in home	rowngun

does gun belong to r	hunt

does r or spouse hunt	hunt1

does r or spouse or partner hunt	news

how often does r read newspaper	tvhours

hours per day watching tv	phone

does r have telephone	coop

r's attitude toward interview	comprend

r's understanding of questions	form

form of split questionnaire asked	fechld

mother working doesn't hurt children	fepresch

preschool kids suffer if mother works	fefam

better for man to work, woman tend home	racdif1

differences due to discrimination	racdif2

differences due to in-born learning ability	racdif3

differences due to lack of education	racdif4

differences due to lack of will	helppoor

should govt improve standard of living?	helpnot

should govt do more or less?	helpsick

should govt help pay for medical care?	helpblk

Should govt aid Black people?	numgiven

number of persons mentioned	race1

race of person number 1	race2

race of person number 2	race3

race of person number 3	race4

race of person number 4	race5

race of person number 5	spouse1

person #1 is r's spouse	spouse2

person #2 is r's spouse	spouse3

person #3 is r's spouse	spouse4

person #4 is r's spouse	spouse5

person #5 is r's spouse	parent1

person #1 is r's parent	parent2

person #2 is r's parent	parent3

person #3 is r's parent	parent4

person #4 is r's parent	parent5

person #5 is r's parent	child1

person #1 is r's child	child2

person #2 is r's child	child3

person #3 is r's child	child4

person #4 is r's child	child5

person #5 is r's child	othfam1

person #1 is r's family member	othfam2

person #2 is r's family member	othfam3

person #3 is r's family member	othfam4

person #4 is r's family member	othfam5

person #5 is r's family member	cowork1

person #1 is r's coworker	cowork2

person #2 is r's coworker	cowork3

person #3 is r's coworker	cowork4

person #4 is r's coworker	cowork5

person #5 is r's coworker	memgrp1

person #1 is member of r's group	memgrp2

person #2 is member of r's group	memgrp3

person #3 is member of r's group	memgrp4

person #4 is member of r's group	memgrp5

person #5 is member of r's group	neighbr1

person #1 is r's neighbor	neighbr2

person #2 is r's neighbor	neighbr3

person #3 is r's neighbor	neighbr4

person #4 is r's neighbor	neighbr5

person #5 is r's neighbor	friend1

person #1 is r's friend	friend2

person #2 is r's friend	friend3

person #3 is r's friend	friend4

person #4 is r's friend	friend5

person #5 is r's friend	advisor1

person #1 is r's advisor	advisor2

person #2 is r's advisor	advisor3

person #3 is r's advisor	advisor4

person #4 is r's advisor	advisor5

person #5 is r's advisor	other1

person #1 is other type of acquaintance	other2

person #2 is other type of acquaintance	other3

person #3 is other type of acquaintance	other4

person #4 is other type of acquaintance	other5

person #5 is other type of acquaintance	talkto1

how often does r talk to person #1?	talkto2

how often does r talk to person #2?	talkto3

how often does r talk to person #3?	talkto4

how often does r talk to person #4?	talkto5

how often does r talk to person #5?	god

r's confidence in the existence of god	reborn

has r ever had a 'born again' experience	wlthwhts

rich - poor	wlthblks

rich - poor	wlthhsps

rich - poor	workwhts

hard working - lazy	workblks

hard working - lazy	workhsps

hard working -lazy	liveblks

neighborhood half black	marblk

close relative marry Black person	marasian

close relative marry asian	marhisp

close relative marry hispanic	marwht

r favor close relative marrying white person	discaff

whites hurt by aff. action	yousup

how many people do you supervise directly	spwksup

does spouse supervise anyone	united

americans united or divided on values	impfinan

imp to r of financial security	impmar

imp to r of being married	impkids

imp to r of having children	impgod

imp to r of having faith in god	impthngs

imp to r of having nice things	impcultr

imp to r of being cultured	impjob

imp to r of having a fulfilling job	impself

imp to r of being self-sufficient	fejobaff

for or against preferential hiring of women	refpromo

refuse a job promotion?	refmorwk

take on additional paid work?	workless

cut back on your work?	noathome

miss a family occasion or holiday?	nonurse

unable to care for a sick child-relative?	discaffm

a man won't get a job or promotion	discaffw

a woman won't get a job or promotion	fehire

should hire and promote women	hapfeel

felt happy?	ashamed

felt ashamed of something you'd done?	fearful

felt fearful about something	angry

felt angry at someone?	proud

felt proud of something you'd done?	vigversn

vignette version number	seriousp

how serious x s situation is	charactr

how likely x s situation caused by bad character	imbalnce

how likely x s situation caused by imbalance in brain	wayraise

how likely x s situation caused by the way raised	stresses

how likely x s situation caused by stress	genetics

how likely x s situation caused by genetic problem	godswill

how likely x s situation caused by god s will	upsdowns

how likely x is experiencing normal ups-and-downs of life	breakdwn

how likely x is experiencing nervous breakdown	mentlill

how likely x is experiencing mental illness	physill

how likely x is experiencing physical illness	viglabel

how likely x is experiencing vignettes	dectreat

how able is x to decide receive treatment	decmoney

how able is x to decide managing own money	imprvown

how likely x s situation will improve on its own	imprvtrt

how likely x s situation will improve with treatment	vignei

how willing r would have x as a neighbor	vigsoc

how willing r would spend time socializing with x	vigfrnd

how willing r would make friends with x	vigwork

how willing r would work closely with x on a job	viggrp

how willing r would have a group home for x	vigmar

how willing r would have x marry in r's family	hurtoth

how likely x violent toward other people	hurtself

how likely x violent toward oneself	tlkfam

x should talk to family and friends	tlkclrgy

x should talk to religious leader	meddoc

x should go to general medical doctor for help	mentldoc

x should go to psychiatrist for help	mentloth

x should go to other mental health professional for help	sphealer

x should go to spiritual healer for help	selfhelp

x should join a self-group with similar problem	otcmed

x should take non-prescription medication	rxmed

x should take prescription medication	mentlhos

x should check into a mental hospital	dofirst

here are the things you said ... should do... interviewer: read list	mustdoc

x should be forced to be examined at a clinic by law	mustmed

x should be forced to take prescribed medication by law	musthosp

x should be admitted to be hospitalized for treatment by law	dangrslf

x should be admitted to be hospitalized for treatment if dangerous to x	dangroth

x should be admitted to be hospitalized for treatment if dangerous to others	knwmhosp

know anyone who was in a hospital	knwpatnt

known anyone seeing a psychologist etc?	fammhneg

r's family's negative attitudes about mh problems	othmhneg

r's other acquaintences negative attitudes about mh problems	diagnosd

r ever diagnosed with mh problem	mhtreatd

r ever treated for mh problem	relpersn

r consider self a religious person	sprtprsn

r consider self a spiritual person	aimofart

aim of art	othlang

can r speak language other than english/spanish	othlang1

what other languages does r speak	othlang2

what other languages does r speak	spklang

how well does r speak other language	letinhsp

immigration from latin america should be increased	letinasn

immigration from asia should be increased	compuse

r use computer	webmob

r uses home internet through mobile device	huclean

r hu interior cleanliness rating	satdemoc

r satisfied with way democracy works in america	disrspct

r is treated with less courtesy or respect than others	poorserv

r recieves poorer service in restaurants or stores	notsmart

people act as if they think r is not smart	afraidof

people act as if they are afraid of r	threaten

r is threatened or harassed	abmoral

r has a moral opposition to abortion	abhelp1

r would help with arrangements for abortion	abhelp2

r would help with paying for abortion	abhelp3

r would help with paying for abortion-related other costs	abhelp4

r would help with emotional support for abortion	numpets

how many pets r's family has	dog

dog	cat

cat	smammal

small mammal (rabbit, gerbil, hamster)	bird

bird	fish

fish	reptile

reptile (turtle, snake, lizard)	othpet

other pet	petfam

how often r considers pet member of family	chldvig

Vignette asked	chldprob

how serious r consider his or her problem	badchar

situation caused by own bad character	chembal

situation by a chemical imbalance in the brain	stressfl

situation caused by stressful circumstances	geneprob

situation caused by a genetic or inherited problem	raised

situation caused by the way he or she was raised	violtv

situation caused by watching violent tv	discipln

situation caused by lacking discipline in the	allergic

situation caused by food or chemical allergies	imprvdis

situation will improve with strict discipline	imprveat

situation will improve with some changes in diet	imprvmed

situation will improve with medical treatment	nextdoor

how willing r move next door to child's family	spendeve

how willing r spend evening with child's family	chldfrnd

how willing r have children make friends with ch	chldsch

how willing r have child in same class of r's ch	adfam

would r get advice from family if child was r's	adtch

would r get advice from teachers child if was r's	addoc

would go to a medical doctor if he was r's child	adcoumed

would r have child on med by counselor's advice	adpsy

would r go to a psychiatrist if child was r's	grwingup

is child experiencing normal ups and downs	mntlill

is child experiencing a mental illness	violpeop

would child do something violent toward other	violself

would child do something violent to own self	numemps

number of employee for the self-employed	wrkslffam

does r work in family business or farm	mhunsure

people like x are unpredictable	mhtrtot2

r knows someone who received treatment for mental health	polint

your interest in politics	poleff11

don't have any say about what the government does	cctv

civil liberties - video surveillance	posslq

does r have marital partner	posslqy

relationship status and cohabitation or not	marcohab

cohabitation status	trcourts

how much trust r has in american courts	goodlife

standard of living of r will improve	marsame

Homosexual couples should have the right to marry one another	meovrwrk

men hurt family when focus on work too much	localnum

number of employees: r's work site	relactiv

how often does r take part in relig activities	befair

how often do you think people take advantage	cantrust

people can be trusted or cant be too careful	religinf

the u.s. would be a better country if religion had less influence	tempgen

greenhouse effect danger to envir	grncon

concerned about environment	tempgen1

greenhouse effect danger to envir	ambornin

how important to have been born in america	amcit

how important to have american citizenship	amchrstn

how important to be a christian	amgovt

how important to respect america's laws etc	amfeel

how important to feel american	amcitizn

agree i would rather be a citizen of america	amshamed

agree there are things make me ashamed	belikeus

agree better if people were more like americans	ambetter

agree america is a better country	ifwrong

agree people should support their country	prouddem

how proud the way democracy works	proudpol

how proud its political influence in the world	proudeco

how proud america's economic achievements	proudspt

how proud its achievements in sports	proudart

how proud its achievements in the arts & lit.	proudhis

how proud its history	imports

america should limit the import	wrldgovt

international bodies should enforce environment	amownway

america should follow its own interests	forland

foreigner's should not be allowed to buy land	amtv

tv should give preference to american films	immcrime

immigrants increase crime rates	immameco

immigrants good for america	immjobs

immigrants take jobs away	letin1

number of immigrants to america nowadays should be	citizen

are you a citizen of america?	amancstr

how important to have american ancestry	intlincs

large intl company damage to local business	letin1a

number of immigrants nowadays should be	immassim

what statement about immigrants matches view	paytaxes

how important never to try to evade taxes	obeylaws

how important always to obey laws	watchgov

how important to keep watch on action of govt	helpusa

how important to help worse off ppl in america	solok

how important: citizens have adequate standard of living	rghtsmin

how important: govt protect right of minorities	polopts

how important: ppl given chance to participate in decsns	oppsegov

how important: citizen engage in acts of civil disobeyed	poleff18

govt do not care much what ppl like r think	corruptn

how widespread corruption is in pub service in america	demtoday

how well democracy work in america	dem10pst

how well did democracy work in america ten yrs ago	dem10fut

how well will democracy work in america in ten yrs	polnews

how often use media to get political news	leftrght

how left or right in politics	partners

how many sex partner's r had in last year	sexsex

sex of sex partner's in last year	sexfreq

frequency of sex during last year	hivtest

have you ever been tested for hiv	hivtest1

in what month and year was your last hiv test	hivtest2

where did you have your last hiv test	sexornt

sexual orientation	realinc

family income in constant $	realrinc

r's income in constant $	coninc

family income in constant dollars	conrinc

respondent income in constant dollars	ethnic

country of family origin	eth1

1st mentioned country of origin	hispanic

hispanic specified	racecen1

what is r's race 1st mention	racecen2

what is r's race 2nd mention	racecen3

what is r's race 3rd mention	uscitzn

is r us citizen	vetyears

years in armed forces	dwelling

type of structure	dwelown

does r own or rent home?	dwelown16

did rs family own or rent home when r was age 16	worda

word a	wordb

word b	wordc

word c	wordd

word d	worde

word e	wordf

word f	wordg

word g	wordh

word h	wordi

word i	wordj

word j	wordsum

number words correct in vocabulary test	hhrace

race of household	hhtype1

household type (condensed)	famgen

number of family generations in household	rplace

r's relationship to household head	rvisitor

is r a visitor?	visitors

number of visitor's in household	dateintv

date of interview	isco08

respondent's occupation, 2010 census & 2008 isco code	paisco08

r's father's occupation, 2010 census & 2008 isco code	maisco08

r's mother's occupation, 2010 census & 2008 isco code	spisco08

r's mothers occupation, 2010 census & 2008 isco code	coisco08

partner's occupation, 2010 census & 2008 isco code	sei10

r's socioeconomic index (2010)	sei10educ

Percentage of some college educ in OCC10 based on ACS 2010	sei10inc

Percentage of $45k+ earners in OCC10 based on ACS 2010	pasei10

r's father's socioeconomic index (2010)	pasei10educ

Percentage of some college educ in PAOCC10 based on ACS 2010	pasei10inc

Percentage of $45k+ earners in PAOCC10 based on ACS 2010	masei10

r's mother's socioeconomic index (2010)	masei10educ

Percentage of some college educ in MAOCC10 based on ACS 2010	masei10inc

Percentage of $45k+ earners in MAOCC10 based on ACS 2010	spsei10

r's spouse's socioeconomic index (2010)	spsei10educ

Percentage of some college educ in SPOCC10 based on ACS 2010	spsei10inc

Percentage of $45k+ earners in SPOCC10 based on ACS 2010	cosei10

partner's socioeconomic index (2010)	cosei10educ

Percentage of some college educ in COOCC10 based on ACS 2010	cosei10inc

Percentage of $45K+ earners in COOCC10 based on ACS 2010	copres10

Partner's occupational prestige score (2010)	copres105plus

partner's occupational prestige score using threshold method (2010)	cohort

year of birth	whoelse1

presence of others: children under six	whoelse2

presence of others: older children	whoelse3

presence of others: spouse partner	whoelse4

presence of others: other relatives	whoelse5

presence of others: other adults	whoelse6

presence of others: no one	intid

Interviewer id	feeused

fee given to get case	feelevel

amount of fees paid	lngthinv

how long was interview	intage

age of interviewer	intethn

race of interviewer	mode

interview done in-person or over the phone	intsex

sex of interviewer	intyrs

years of service as an interviewer at norc	consent

participation/recording consent	adminconsent

consent to possible data linkage	letdie1y

allow incurable patients to die (form 2)	ballot

ballot used for interview	issp

filter for issp cases	sampcode

Sampling error code	sample

sampling frame and method	oversamp

Weights for black oversamples	phase

subsampling: two-phase design.	spanself

if this interview had only been available in english, would you...	spaneng

interviews conducted in spanish or english	hlthstrt

would you say the respondent's health in general is excellent, good, fair, or poor	kidsinhh

does interviewer believe that there are children under the age of 15 present at	respond

estimate of the probability that this household will respond to the survey	incuspop

estimated income status of housing unit	neisafe

how safe interviewer thinks neighborhood is	rlooks

r physical attractiveness rating	rgroomed

r grooming rating	rhlthend

r health rating	vstrat

Variance stratum	vpsu

Variance primary sampling unit	kish

famdif16y

Which of the following best describes your situation?	pawrkslf2

At this job, was [dadfill] an employee, self-employed without employees, or self-employed with employees?	pawrkslffam

Father (or oldest same-sex parent) owned family business or farm	mawrkslf2

At this job, was mother (or youngest same-sex parent) an employee, self-employed without employees, or self-employed with employees?	mawrkslffam

Mother (or youngest same-sex parent) owned family business or farm	ethworld1

From what part of the world did your ancestors come? Europe	ethworld2

From what part of the world did your ancestors come? Africa	ethworld3

From what part of the world did your ancestors come? Asia	ethworld4

From what part of the world did your ancestors come? Middle East	ethworld5

From what part of the world did your ancestors come? South America	ethworld6

From what part of the world did your ancestors come? Central America	ethworld7

From what part of the world did your ancestors come? North America	ethworld8

From what part of the world did your ancestors come? Caribbean	ethworld9

From what part of the world did your ancestors come? Oceania	wrkgovt1

govt employee	wrkgovt2

private employee	spkathy

allow anti-religionist to speak y	libathy

allow anti-religionist to teach y	spkracy

allow anti-religious book in library y	libracy

allow racists book in library y	spkcomy

allow communist to speak y	colcomy

should communist teacher be fired y	libcomy

allow communists book in library y	spkmslmy

allow muslim clergymen preaching hatred of the us y	libmslmy

allow anti-american muslim clergymen's books in library y	polhitoky

ever approve of police striking citizen y	polabusey

citizen said vulgar or obscene things y	polattaky

citizen attacking policeman with fists y	raceacs1

white	raceacs2

black	raceacs3

american indian or alaska native	raceacs4

asian indian	raceacs5

chinese	raceacs6

filipino	raceacs7

japanese	raceacs8

korean	raceacs9

vietnamese	raceacs10

other asian	raceacs14

other pacific islander	raceacs15

some other race	raceacs16

hispanic	abdefectg

strong chance of serious defect (grid on web)	abnomoreg

married–wants no more children (grid on web)	abhlthg

woman's health seriously endangered (grid on web)	abpoorg

low income–cant afford more children (grid on web)	abrapeg

pregnant as result of rape (grid on web)	absingleg

not married (grid on web)	suicide1g

suicide if incurable disease	suicide2g

suicide if bankrupt	suicide4g

suicide if tired of living	maborn

Was R's mother (or youngest same-sex parent) born in this country?	paborn

Was R's father (or oldest same-sex parent) born in this country?	sexbirth1

r's sex assigned at birth (2021)	norelgsp

is r atheist or agnostic	hivafraid

afraid to be around a person with hiv	hivimmrl

people with hiv have participated in immoral activities	hivdscrm

there is a lot of discrimination against people with hiv	ptnrornt

recent partner's sexual orientation	ptnrsxbrth

recent partner's sex at birth	wtssps

person post-stratification weight	wtssnrps

Person post-stratification weight, nonrespondents adjusted	prayerv

bible prayer in public schools (with volunteered response on web)	courtsv

courts dealing with criminals (with volunteered response on web)	discaffwv

a woman won't get a job or promotion (with volunteered response on web)	racopenv

vote on open housing law (with volunteered response on web)	getaheadv

opinion of how people get ahead (with volunteered response on web)	divlawv

divorce laws (with volunteered response on web)	helpfulv

people helpful or looking out for selves (with volunteered response on web)	fairv

people fair or try to take advantage (with volunteered resposne on web)	trustv

can people be trusted (with volunteered response on web)	grassv

should marijuana be made legal (with volunteered response on web)	relitenv

strength of affiliation (with volunteered response on web)	biblev

feelings about the bible (with volunteered response on web)	postlifev

belief in life after death (with volunteered response on web)	kidssolv

r's kids living standard compared to r (with volunteered response on web)	fepolv

women not suited for politics (with volunteered response on web)	prayernv

bible prayer in public schools (no volunteered response on web)	courtsnv

courts dealing with criminals (no volunteered response on web)	discaffwnv

a woman won't get a job or promotion (no volunteered response on web)	racopennv

vote on open housing law (no volunteered response on web)	getaheadnv

opinion of how people get ahead (no volunteered response on web)	divlawnv

divorce laws (no volunteered response on web)	helpfulnv

people helpful or looking out for selves (no volunteered response on web)	fairnv

people fair or try to take advantage (no volunteered response on web)	trustnv

can people be trusted (no volunteered response on web)	grassnv

should marijuana be made legal (no volunteered response on web)	relitennv

strength of affiliation (no volunteered response on web)	biblenv

feelings about the bible (no volunteered response on web)	postlifenv

belief in life after death (no volunteered response on web)	kidssolnv

r's kids living standard compared to r (no volunteered response on web)	fepolnv

women not suited for politics (no volunteered response on web)	abanyg

abortion if woman wants for any reason (grid on web)	fileversion

NA	spwrkslf2

SPOUSE SELF-EMPLOYED, WITH EMPLOYEES, OR WORK FOR SOMEONE ELSE	spwrkslffam

SPOUSE WORK AT FAMILY BUSINESS OR ON FAMILY FARM	childsinhh

TOTAL CHILDREN IN HH (HOUSEHOLD COMPOSITION VERSION)	adultsinhh

TOTAL ADULTS IN HH (HOUSEHOLD COMPOSITION VERSION)	whatsp2

How would you classify the place where your ...? Was this mainly...?	cowrkslf2

PARTNER SELF-EMPLOYED, WITH EMPLOYEES, OR WORK FOR SOMEONE ELSE	cowrkslffam

PARTNER WORK AT FAMILY BUSINESS OR ON FAMILY FARM	cowksup

DOES PARTNER SUPERVISE ANYONE	whatco2

CLASSIFY PLACE WHERE PARTNERS WORKS	whatpa2

How would you classify the place where father (or oldest same-sex parent) worked? Was this mainly...?	whatma2

How would you classify the place where mother (or youngest same-sex parent) worked? Was this mainly...?	whatslf2

How would you classify the place where you ...? Was this mainly...?	racerank1

FIRST RACE SELECTED	racerank2

SECOND RACE SELECTED	racerank3

THIRD RACE SELECTED	adoptus

IMMIGRANTS LIVING IN US SHOULD ADOPT US CULTURE	immfate

FATE OF ILLEGAL IMMIGRANTS	xmoviey

SEEN PORNOGRAPHY IN PAST YEAR	vote20

REMEMBER IF VOTED IN 2020 ELECTION	pres20

VOTED TRUMP OR BIDEN	if20who

WHO YOU WOULD HAVE VOTED FOR IN 2020	wordk

WORD K	wordl

WORD L	wordn

WORD N	svyenjoy

ENJOY TAKING SURVEYS	svyid1

WILLINGNESS TO USE VIDEO CHAT	svyid2

WILLINGNES FOR IN-PERSON SURVEY	gestate

ABORTION LEGAL PAST 4.5 MONTHS	nextstatus

GSS NEXT FOLLOW-ON STATUS	racdif1x

DIFFERENCES DUE TO DISCRIMINATION (LONGER INTRO)	racdif2x

DIFFERENCES DUE TO IN-BORN LEARNING ABILITY	racdif3x

DIFFERENCES DUE TO LACK OF EDUCATION	racdif4x

DIFFERENCES DUE TO LACK OF WILL	hompop_exp

TOTAL PEOPLE IN HH (EXPERIMENTAL VERSION)	modesequence

INITIAL MODE OF CONTACT FOR RESPONDENT	totalincentive

totalincentive	babies_exp

KIDS 0-6 IN HH(EXPERIMENTAL VERSION)	preteen_exp

KIDS 7-12 IN HH(EXPERIMENTAL VERSION)	teens_exp

KIDS 13-17 IN HH(EXPERIMENTAL VERSION)	adults_exp

ADULTS 18+ IN HH(EXPERIMENTAL VERSION)	prnt

R IS A PARENT IN SOME CAPACITY IN HH	childs_exp

TOTAL CHILDREN UNDER 18 IN HH (EXPERIMENTAL VERISON)	famgen_exp

NUMBER OF FAMILY GENERATIONS IN HOUSEHOLD (EXPERIMENTAL VERSION)	hhtype1_exp

HOUSEHOLD TYPE (CONDENSED) (EXPERIMENTAL VERSION)	batch

SAMPLE BATCH	subsamprate

Subsampling rate corresponding to two-phase design	wtssps_next

POSTSTRATIFICATION WEIGHT FOR ANALYZING GSS NEXT FOLLOW-ON	wtssnrps_next

POSTSTRATIFICATION AND NONRESPONSE WEIGHT FOR ANALYZING GSS NEXT FOLLOW-ON	nexteligible

GSS NEXT FOLLOW-ON ELIGIBLE	pawrkgrw

Did father (or oldest same-sex parent) ever work for pay for as long as a year while you were growing up?	conlabory

Confidence in labor unions	racdif5

differences due mainly to unbringing	marsamey

Same-sex couples should have the right to marry one another	norelgsp16

R no religion age 16	spnorelgsp

Spouse no religion	mobiledata

HOUSEHOLD HAS ACCESS TO INTERNET USING CELLULAR DATA PLAN	broadband

HOUSEHOLD HAS ACCESS TO INTERNET USING BROADBAND INTERNET SERVICE	rateincumb24

RATING OF JOE BIDEN	ratechall124

RATING OF DONALD TRUMP	ratechall224

RATING OF KAMALA HARRIS	ratechall324

RATING OF NIKKI HALEY	ratedemp

RATING OF DEMOCRATIC PARTY	raterepp

RATING OF REPUBLICAN PARTY	ratemaga

RATING OF MAGA REPUBLICANS	lkelyvot

LIKELIHOOD TO VOTE IN 2024 ELECTION	whovote24a

VOTING INTENTION IN 2024 ELECTION (POST JULY 25 2024)	econstat

WHAT DO YOU THINK ABOUT THE STATE OF THE ECONOMY THESE DAYS IN THE UNITED STATES?	polintrst

INTERESTED IN CAMPAIGNS THIS YEAR	intrnetuse

Frequency of internet use over past 12 months	nouseint1

No internet use over past 12 months due to unavaliblity	nouseint2

No internet use over past 12 months due to poor quality of connection	nouseint3

No internet use over past 12 months due to cost	nouseint4

No internet use over past 12 months due to security and privacy	nouseint5

No internet use over past 12 months due to lack of utility/importance	nouseint6

No internet use over past 12 months due to age	nouseint7

No internet use over past 12 months due to usability	nouseint8

No internet use over past 12 months due to lack of desire	nouseint9

No internet use over past 12 months due to literacy	nouseint10

No internet use over past 12 months due to lack of time	nouseint11

No internet use over past 12 months due to government restriction	nouseint12

No internet use over past 12 months due to other reasons	intaccprxy

How often R asked someone to use the internet on their behalf	prxyuse

Likelihood R ask someone to use the internet on their behalf in near future	ftruse

Likelihood R uses internet in next 12 months	inthome

Frequency R has home internet access	intprfssnl

Frequency R has work/school internet access	intpub

Frequency R has internet access in public	intbhlf

Frequency R was asked to use internet for someone else	unplug

Amount R reduced time spent on internet in last 12 months	unplgrsn

Reason R reduced time spent on internet	intskill

R's skill at using the internet	intsurf

R's skill at online searches	intliteracy

R's knowledge of online information discretion	intapps

R's skill at learning new software	intcomm

How often R uses internet for communication	intshare

How often R uses internet for sharing photos/videos	intsearch

How often R uses internet for search	intgame

How often R uses internet for video games	intstream

How often R uses internet for music/video	intfncl

How often R uses internet for financial transactions	intmeet

R more comfortable meeting people online?	intlnly

R would feel lonely without internet?	inttrig

How often R reads political opinions they disagree with online	intnews

How often R accesses non-routine political news sources	intcnfrm

How often R tries to confirm information with other sources	intrust

How much R trusts people they've only met online	intpolview

Trend of online contact with politically similar	intoppo

Trend of online contact with politically different	intviews

How often R talks about politics online	techesy

Does technology make our lives easier?	harmgood1

Does technology do more harm than good?	nextgen1

Does technology provide more opportunities?	intscam

Has R or someone they know been a victim of online scam	inthrss

Has R or someone they know been harassed online	polnewsfrom

Where R gets political news	smnews

Reliability of news on social media	tvnews1

Reliability of news on TV	papernews

Reliability of newspapers	radionews

Reliability of news on podcasts, radio	webnews

Reliability of news on web	modepet

Has R signed a petition in last 12 months	modeprot

Has R participated in a protest in last 12 months	modelobby

Has R contacted a politician in last 12 months	modeorgprot

Has R organized a protest in last 12 months	modeorg

Has R belonged to an activist group in last 12 months	trppl

Trust in others, 10 point scale	trcong

Trust in congress, 10 point scale	dataprot

Responsibility for internet data protection	emonitor1

Should government be able to monitor online communication	infodeal

Would R provide info for discounts, free products	infoprofit

Would R provide info companies could profit from	gendtech

Are women or men benefiting more from digital tech	agetech

Are older or younger people benefiting more from digital tech	urbantech

Are cities or countryside benefiting more from digital tech	eductech

Are educated or uneducated benefiting more from digital tech	classtech

Are poor or rich benefiting more from digital tech	aiworry

R worried that AI will take over many jobs done by humans	aimed

R's comfort with medical operation by robot	aidrive

R's comfort with driverless car	owndecns

Best if every nation had its own country	brneffrt

Possible to become American if born elsewhere	immcrime1

Immigrants increase crime rates	immjobs1

Immigrants take jobs away	immideas1

Immigrants bring new ideas and cultures	immwlfare

People born in America should be given preference over immigrants when it comes to jobs, housing, or health care	conscrp

Importance of military service	incgap1

Government's responsibility to reduce economic inequality	leftrght1

How left or right in politics	servpub

How much the public service serves people	fulldem

The people, not politicians, should make important decisions	poltalk

Politicians talk too much and don't take action	conpow

Conflict between ordinary people and those in power	bizmoney

Business leaders make money at expense of ordinary people	experts

Better to rely on opinions of experts than ordinary people	biasnews

News coverage about politics in America is biased	usmca1

R's familiarity with USMCA	usmca2

Does the US benefit from USMCA	morextrm

Everyone is like X but X's case is just a bit more extreme and ongoing	bullied

Situation caused by having been bullied	smedia

Situation caused by exposure to social media	imprvown1

How likely child X situation will improve on its own	mustdoc1

X's parents should be forced by law to have X examined at a clinic	mustmed1

X's parents should be forced by law to have X take prescribed medication	musthosp1

X's parents should be forced by law to admit X to be hospitalized for treatment	physill1

How likely child X is experiencing physical illness	bullpeer

Would X bully peers	rumopeer

Would X spread rumors about peers	fighpeer

Would X get into fights with peers	schshoot

Would X carry out a school shooting	cutself

Would X intentionally cut self	negself

Would X intentionally neglect own health	foodalc

Would X engage in excessive and uncontrolled food or alcohol consumption	risklife

Would X at risk for taking their own life	takelife

Would X actually take own life	dsndntpay

Reparations for descendants of slaves	acpttg

Accepting people	chngdocs

Changing official documents	feovrwrk

Family life suffers when women work	sexornt1

Sexual orientation (reask, self-administered)	sexpaid

Ever had sex with a person you paid or who paid you for sex	nocondom

Had vaginal or anal sex without a condom in past 12 months	nodocidu

Ever injected any drugs other than those prescribed for you	femself1

How feminine R sees self (0 to 6 scale)	mascself1

How masculine R sees self (0 to 6 scale)	femsee1

How feminine others see R (0 to 6 scale)	mascsee1

How masculine others see R (0 to 6 scale)	devtype

Are you taking this survey on a...	family16sex

Sex of two parents living with when 16 yrs old	racdif5x

differences due mainly to upbringing (longer intro)	reltrad

Religious tradition	reltrad16

Religious tradition age 16	conorelgsp

Partner no religion	sexnow2

R's sex now (3 category)	mighist

When R came to U.S.	migage

Age when R came to U.S.	whovote24

VOTING INTENTION IN 2024 ELECTION (PRE JULY 25 2024)	whovotets

VOTING INTENTION - TIME SERIES	voteprob

VOTING PROBLEM IN US ELECTIONS	rateincumbts

RATING OF INCUMBENT PRESIDENT - TIME SERIES	ratechall1ts

RATING OF PRESIDENTIAL CHALLENGER 1 - TIME SERIES	ratechall2ts

RATING OF PRESIDENTIAL CHALLENGER 2 - TIME SERIES	ratechall3ts

RATING OF PRESIDENTIAL CHALLENGER 3 - TIME SERIES	aneseligible

ANES FOLLOW-ON ELIGIBLE	anespre_status

ANES PRE-ELECTION SURVEY STATUS	anespost_status

ANES POST-ELECTION SURVEY STATUS	chldfam_asn

Percent of family during childhood who were Asian	chldfam_blk

Percent of family during childhood who were Black	chldfam_hisp

Percent of family during childhood who were Latino	chldfam_wht

Percent of family during childhood who were White	chldfam_oth

Percent of family during childhood who were some other race	chldadlt_asn

Percent of adults you knew (friends, care takers, neighbors) during childhood who were Asian	chldadlt_blk

Percent of adults you knew (friends, care takers, neighbors) during childhood who were Black	chldadlt_hisp

Percent of adults you knew (friends, care takers, neighbors) during childhood who were Latino	chldadlt_wht

Percent of adults you knew (friends, care takers, neighbors) during childhood who were White	chldadlt_oth

Percent of adults you knew (friends, care takers, neighbors) during childhood who were some other race	chldchld_asn

Percent of children you knew (friends, classmates) during childhood who were Asian	chldchld_blk

Percent of children you knew (friends, classmates) during childhood who were Black	chldchld_hisp

Percent of children you knew (friends, classmates) during childhood who were Latino	chldchld_wht

Percent of children you knew (friends, classmates) during childhood who were White	chldchld_oth

Percent of children you knew (friends, classmates) during childhood who were some other race	chldshop_asn

Percent of people who shopped at the supermarket during childhood who were Asian	chldshop_blk

Percent of people who shopped at the supermarket during childhood who were Black	chldshop_hisp

Percent of people who shopped at the supermarket during childhood who were Latino	chldshop_wht

Percent of people who shopped at the supermarket during childhood who were White	chldshop_oth

Percent of people who shopped at the supermarket during childhood who were some other race	ptnrsxnow1

Recent partner's sex now (3 category)	perla

R skin tone via interviewer	petsseq

Position of pet questions in questionnaire	credrep

R credit report	credchoice

R's credit score has to do with choices or things out of your control	credam

Americans' credit score has to do with choices or things out of their control	credland

Use credit report - landlords	credemp

Use credit report - employers	credcar

Use credit report - car insurers	chldidel_next

Ideal number of children - follow-on version	chldpref_next

R ideal number of children - follow-on version	knwsexor

R knows anyone who is gay, lesbian, or bisexual	happy_next

General happiness - follow-on version	perla_self

R self-reported skin tone	topprob1_next

Most important issue for America - follow on version	topprob2_hlth

Next important issue (health care) - follow-on version	topprob2_educ

Next important issue (education) - follow-on version	topprob2_crim

Next important issue (crime) - follow-on version	topprob2_envr

Next important issue (environment) - follow-on version	topprob2_immg

Next important issue (immigration) - follow-on version	topprob2_econ

Next important issue (economy) - follow-on version	topprob2_terr

Next important issue (terrorism) - follow-on version	topprob2_povt

Next important issue (poverty) - follow-on version	topprob2_none

Next important issue (none of these) - follow-on version	hhtype1_next

Household type (condensed) - follow-on version	famgen_next

Number of family generations in household - follow-on version	dateintv_next

Date of GSS Next interview	wordsum7

number words correct in vocabulary test (7 common items)	wordsumold

number words correct in vocabulary test (original items)	wordsumnew

number words correct in vocabulary test (experiment items)

Source

Data: https://kjhealy.github.io/gssr/

Codebook: https://gss.norc.org/content/dam/gss/get-documentation/pdf/codebook/GSS%202024%20Codebook%20R3.pdf