Uniform Crime Reports, 2015 (County-Level)

This subset of data comes from one iteration of the Uniform Crime Reporting Program, administered in 2015. These data were collected by the Federal Bureau of Investigation under the United States Department of Justice. While the original data cover every reported crime event that took place in 2015, these data are aggregated to the county level. Additionally, these data are combined with (a subset of) county-level demographic data from the 2005-2009 (5-year estimates) iteration of the American Community Survey. Information about the data set can be found in the UCR2015 Codebook at: https://burrelvannjr.com/docs/UCR2015_Codebook.pdf.

Usage

UCR2015

Format

A data frame with 3108 observations and 102 variables.

id

State and County Identifier	statefips

FIPS Code for State	countyfips

FIPS Code for County	state

State Name	county

County Name	totalpop

Total County Population	pct_unemp

Percent of Total County Population who are Unemployed	pct_homeowners

Percent of Total County Population who are Homeowners	pct_college

Percent of Total County Population who are over 25 years old and hold a Bachelor's Degree	med_fam_inc

Median Family Income (in Thousands of Dollars)	pop_density

Population Density (Population over Land Area in County)	pct_poverty

Percent of Total County Population who are below the Poverty Line	pct_white

Percent of Total County Population who are White	pct_black

Percent of Total County Population who are Black	pct_latino

Percent of Total County Population who are Latinx/e/a/o	income_inequality

Gini Coefficient of Income Inequality – The distribution of income across the county population. High scores indicate greater inequality, with high-income individuals receiving much larger percentages of the total income made in the county.	rape

Forcible rape (Count)	robbery

Robbery (Count)	agg_assault

Aggravated assault (Count)	burglary

Burglary-breaking or entering (Count)	larceny

Larceny-theft (not motor vehicles) (Count)	mv_theft

Motor vehicle theft (Count)	other_assault

Other assaults (Count)	arson

Arson (Count)	forgery

Forgery and counterfeiting (Count)	fraud

Fraud (Count)	embezzlement

Embezzlement (Count)	stolen_property

Stolen property-buy, receive, poss. (Count)	vandalism

Vandalism (Count)	weapons

Weapons-carry, posses, etc. (Count)	sex_offense

Sex offenses (not rape or prostitution) (Count)	drug_abuse

Total drug abuse violations (Count)	drug_sale

Sale/manufacture (subtotal) (Count)	drug_possession

Possession (subtotal) (Count)	drug_sale_coke

Sale/mfg-Opium, coke, and their derivatives (Count)	drug_sale_mj

Sale/mfg-Marijuana (Count)	drug_possession_coke

Possession-Opium, coke, and their derivatives (Count)	drug_possession_mj

Possession-Marijuana (Count)	drug_possession_narc

Possession-Truly addicting synthetic narcotics (Count)	drug_possession_other

Possession-Other dangerous non-narc drugs (Count)	domestic_offenses

Offenses against family and children (Count)	dui

Driving under the influence (Count)	liquor_violation

Liquor laws (Count)	disorderly_conduct

Disorderly conduct (Count)	other_nontraffic_violation

All other non-traffic offenses (Count)	murder

Murder and non-negligent manslaughter (Count)	drug_sale_other

Sale/mfg-Other dangerous non-narc drugs (Count)	prostitution

Prostitution and commercialized vice (Count)	drug_sale_narc

Sale/mfg-Truly addicting synthetic narcotics (Count)	vagrancy

Vagrancy (Count)	drunkenness

Drunkenness (Count)	curfew_loitering

Curfew and loitering violations (Count)	runaway

Runaways (Count)	manslaughter_negligence

Manslaughter by negligence (Count)	gambling_all

Gambling (total) (Count)	suspicion

Suspicion (Count)	gambling_bookmaking

Bookmaking (horse and sports) (Count)	gambling_other

All other gambling (Count)	gambling_lottery

Number and lottery (Count)	rape_pct

Forcible rape (as percent of total county population)	robbery_pct

Robbery (as percent of total county population)	agg_assault_pct

Aggravated assault (as percent of total county population)	burglary_pct

Burglary-breaking or entering (as percent of total county population)	larceny_pct

Larceny-theft (not motor vehicles) (as percent of total county population)	mv_theft_pct

Motor vehicle theft (as percent of total county population)	other_assault_pct

Other assaults (as percent of total county population)	arson_pct

Arson (as percent of total county population)	forgery_pct

Forgery and counterfeiting (as percent of total county population)	fraud_pct

Fraud (as percent of total county population)	embezzlement_pct

Embezzlement (as percent of total county population)	stolen_property_pct

Stolen property-buy, receive, poss. (as percent of total county population)	vandalism_pct

Vandalism (as percent of total county population)	weapons_pct

Weapons-carry, posses, etc. (as percent of total county population)	sex_offense_pct

Sex offenses (not rape or prostitution) (as percent of total county population)	drug_abuse_pct

Total drug abuse violations (as percent of total county population)	drug_sale_pct

Sale/manufacture (subtotal) (as percent of total county population)	drug_possession_pct

Possession (subtotal) (as percent of total county population)	drug_sale_coke_pct

Sale/mfg-Opium, coke, and their derivatives (as percent of total county population)	drug_sale_mj_pct

Sale/mfg-Marijuana (as percent of total county population)	drug_possession_coke_pct

Possession-Opium, coke, and their derivatives (as percent of total county population)	drug_possession_mj_pct

Possession-Marijuana (as percent of total county population)	drug_possession_narc_pct

Possession-Truly addicting synthetic narcotics (as percent of total county population)	drug_possession_other_pct

Possession-Other dangerous non-narc drugs (as percent of total county population)	domestic_offenses_pct

Offenses against family and children (as percent of total county population)	dui_pct

Driving under the influence (as percent of total county population)	liquor_violation_pct

Liquor laws (as percent of total county population)	disorderly_conduct_pct

Disorderly conduct (as percent of total county population)	other_nontraffic_violation_pct

All other non-traffic offenses (as percent of total county population)	murder_pct

Murder and non-negligent manslaughter (as percent of total county population)	drug_sale_other_pct

Sale/mfg-Other dangerous non-narc drugs (as percent of total county population)	prostitution_pct

Prostitution and commercialized vice (as percent of total county population)	drug_sale_narc_pct

Sale/mfg-Truly addicting synthetic narcotics (as percent of total county population)	vagrancy_pct

Vagrancy (as percent of total county population)	drunkenness_pct

Drunkenness (as percent of total county population)	curfew_loitering_pct

Curfew and loitering violations (as percent of total county population)	runaway_pct

Runaways (as percent of total county population)	manslaughter_negligence_pct

Manslaughter by negligence (as percent of total county population)	gambling_all_pct

Gambling (total) (as percent of total county population)	suspicion_pct

Suspicion (as percent of total county population)	gambling_bookmaking_pct

Bookmaking (horse and sports) (as percent of total county population)	gambling_other_pct

All other gambling (as percent of total county population)	gambling_lottery_pct

Number and lottery (as percent of total county population)

Source

Data: https://www.icpsr.umich.edu/web/NACJD/studies/36794 and https://data.census.gov/app/mdat/ACSPUMS1Y2023

Codebook: https://burrelvannjr.com/docs/UCR2015_Codebook.pdf