Well-Being and Basic Needs Survey, 2019 (Individual-Level)

This subset of data comes from one iteration of the Well-Being and Basic Needs Survey, administered in 2019. These data were collected by the Urban Institute. Information about the data set can be found in the WBBN2019 Codebook at: https://burrelvannjr.com/docs/WBBN2019_Codebook.pdf.

Usage

WBBN2019

Format

A data frame with 7694 observations and 23 variables.

subsidized_housing

Is your household paying lower rent because the federal, state, or local government is paying part of the cost?	food_last

Food did not last	nervous

During the past 30 days, about how often did you feel: nervous?	hopeless

During the past 30 days, about how often did you feel: hopeless?	restless

During the past 30 days, about how often did you feel: restless or fidgety?	no_cheer

During the past 30 days, about how often did you feel: so sad that nothing could cheer you up	worthless

During the past 30 days, about how often did you feel: worthless?	insured

Thinking about your health insurance coverage over the past 12 months, how many months were you insured?	med_notafford

Thinking about your health care experiences over the past 12 months, was there any time when you needed medical care but did not get it because you couldn't afford it?	working

Are you currently working for pay or self-employed?	unexp_400

How confident are you that you could come up with $400 if an unexpected expense arose within the next month?	educ

Education level	race_eth

Race/ethnicity	sex_gender

Sex/Gender	head_household

Head of Household?	internet

Internet access	children_in_house

Number of children age 0-18 in household	food_insecure

Household was food insecure in past 12 months	utility_suspend

Gas or electric company turned off service or oil company would not deliver in oil past 12 months	utility_problems_paying

Household was not able to pay full amount of gas, oil, or electricity bills in past 12 months	mortgage_cost

How much is the regular monthly payment on this property, including mortgage payments, second mortgage or home equity loan payments, real estate taxes, insurance, and condominium fees?	rent_cost

What is the monthly rent for the place where you live?	electricity_cost

In a typical month, what is the total cost of electricity, gas, and any other fuel used in the place where you live?

Source

Data: https://www.icpsr.umich.edu/web/ICPSR/studies/38044

Codebook: https://burrelvannjr.com/docs/WBBN2019_Codebook.pdf